Begin Converting Styling Functions to Format Selected Text Only
The beginning of my attempts to combine singamandeep's Font Toolbar that formats all the text on the text widget with my Text Editor's ability to apply text styles to selected text only and my .DocX experiments to save formatted text to a .docx file format.  The hope is to create a working MS Wordpad Clone with Python 3.x using the Tkinter library.