Removed Every Instance of Font Size Represented by Str
When I highlight a passage of text and change its font size, the font size combobox only shows the first digit of the new size when I click on it again.  

Text that is made size 10 shows 1 on the combobox.  Text that is made size 20 show 2 on the combobox.  Text that is made size 30 show 3 on the combobox.  

Nothing I do seems to changes this, but otherwise, the code seem to be working as intended.